+++ src/html/undefined-elements.ts
@@ -41,15 +41,24 @@ export class UndefinedElements extends HtmlRule {
         // HACK. Filed as https://github.com/Polymer/polymer-analyzer/issues/507
         continue;
       }
+      // TODO(rictic): ASTNodes should always exist for element references, and
+      //   it should always be possible to get their start tags, but we saw some
+      //   errors where the source range was undefined. Needs investigation.
+      if (!ref.astNode) {
+        continue;
+      }
       const el = document.getById(
           'element', ref.tagName, {imported: true, externalPackages: true});
 
       if (el.size === 0) {
+        const sourceRange = parsedDocument.sourceRangeForStartTag(ref.astNode);
+        if (!sourceRange) {
+          continue;
+        }
         warnings.push({
           code: 'undefined-elements',
           message: `The element ${ref.tagName} is not defined`,
-          severity: Severity.WARNING,
-          sourceRange: parsedDocument.sourceRangeForStartTag(ref.astNode)!
+          severity: Severity.WARNING, sourceRange
         });
       }
     }